finished Data loader started trainer

+++ src/data_loader.py
@@ -4,7 +4,6 @@ import utils
 import os
 import shutil
 import time
-import psutil
 import collections
 
 
@@ -13,48 +12,56 @@ from torch.utils.data import Dataset
 from tabulate import tabulate
 import cv2
 from tqdm import tqdm
+import numpy as np
 
 
 DataFolder = collections.namedtuple('DataFolder', ['path', 'elements'])
-FramePath = collections.namedtuple('FramePath', ['path', 'id'])
 
 class backbone_dataset(Dataset):
     def __init__(self):
         self.cache = None
         self.data_folders = []
         self.bytes_that_I_need = None
-        self.slow_memory_path = {}
-
+        self.max_chars_data_folder_name_lenght = None
+        self.ram_for_stooring_data_retrival_information = None
+        self.rank = int(os.environ['RANK'])
+        # The real data corpus:
+        self.data_path = None
+        self.data_id = None
+        self.data_should_be_in_cache = None
+        self.data_already_in_cache = None
+        # end
 
         self.set_up_cache()
         self.total_number_of_frames = self.check_dataset_and_get_data_folders()
         self.bytes_per_frame = self.get_one_frame_memory_size()
-        self.print_dataset_summary_table()
+        if self.rank == 0:
+            self.print_dataset_summary_table()
         self.last_frame_in_cache = self.get_last_frames_to_put_in_cache()
-        self.print_memory_summary_table()
-        self.load_data_in_cache()
-
-
-    def __len__(self):
-        pass
-    
-    def __getitem__(self, idx):
-        pass
+        self.make_data()
+        if self.rank == 0:
+            self.print_memory_summary_table()
+        del self.data_folders
 
     def close(self):
         self.clean_cache()
 
     def set_up_cache(self):
         if "TMPDIR" in os.environ:
-            print("Using TMPDIR as CACHE!")
+            if self.rank == 0:
+                print("Using TMPDIR as CACHE!")
             config.JOB_TMP_DIR = os.path.join(os.environ["TMPDIR"], config.JOB_TMP_DIR_NAME)
         else:
-            print("Not using TMPDIR as CACHE!")
+            if self.rank == 0:
+                print("Not using TMPDIR as CACHE!")
             config.JOB_TMP_DIR = os.path.join(os.getcwd(), config.JOB_TMP_DIR_NAME)
-        if os.path.isdir(config.JOB_TMP_DIR):
-            shutil.rmtree(config.JOB_TMP_DIR)
-        os.mkdir(config.JOB_TMP_DIR)
-        print(f"CACHE PATH is '{config.JOB_TMP_DIR}' [size_limit={int(config.CACHE_SIZE_LIMIT/1e9)} GB]")
+        if self.rank == 0:
+            if os.path.isdir(config.JOB_TMP_DIR):
+                shutil.rmtree(config.JOB_TMP_DIR)
+                os.mkdir(config.JOB_TMP_DIR)
+            else:
+                os.mkdir(config.JOB_TMP_DIR)
+            print(f"CACHE PATH is '{config.JOB_TMP_DIR}' [size_limit={int(config.CACHE_SIZE_LIMIT/1e9)} GB]")
         self.cache = Cache(directory=config.JOB_TMP_DIR, size_limit=config.CACHE_SIZE_LIMIT)
 
     def clean_cache(self):
@@ -67,15 +74,15 @@ class backbone_dataset(Dataset):
             raise Exception(utils.color_error_string("No Dataset folder found!"))
         
         def check_that_all_sensors_have_the_same_ammount_of_frame(data_folder_path):
-            number_of_frames = None
-            for root, dirs, files in os.walk(data_folder_path, topdown=False):
-                if root == data_folder_path:
-                    continue
+            number_of_frames = None        
+            for data_subfolder in config.DATASET_FOLDER_STRUCT:
+                subfolder_to_check = os.path.join(data_folder_path, data_subfolder[0])
+                files = os.listdir(subfolder_to_check)
                 if number_of_frames is None:
                     number_of_frames = len(files)
                 else:
                     if len(files) != number_of_frames:
-                        raise Exception(utils.color_error_string(f"The folder '{root}' has {len(files)} instead of {number_of_frames}!"))
+                        raise Exception(utils.color_error_string(f"The folder '{root}' has {len(files)} files instead of {number_of_frames}!"))
             return number_of_frames
 
         total_number_of_frames = 0
@@ -91,7 +98,16 @@ class backbone_dataset(Dataset):
                 continue
             elements = check_that_all_sensors_have_the_same_ammount_of_frame(root)
             total_number_of_frames += elements
-            self.data_folders.append(DataFolder(root, elements))
+            # In the following I'm assuming that the folder containing all the data
+            # (ex: 25_02_2024_21:20:57) is a direct child of config.DATASET_PATH!
+            head, tail = os.path.split(root)
+            if os.path.normpath(head) != os.path.normpath(config.DATASET_PATH):
+                raise Exception(utils.color_error_string(f"The folder '{root}' is a valid dataset subfolder but it's not a direct child of '{config.DATASET_PATH}'!"))               
+            self.data_folders.append(DataFolder(tail, elements))
+            if self.max_chars_data_folder_name_lenght is None:
+                self.max_chars_data_folder_name_lenght = len(tail)
+            elif len(tail) > self.max_chars_data_folder_name_lenght:
+                self.max_chars_data_folder_name_lenght = len(tail)
         return total_number_of_frames
             
     def get_one_frame_memory_size(self):
@@ -100,7 +116,7 @@ class backbone_dataset(Dataset):
             return None
         a_data_folder = self.data_folders[0]
         total_size = 0
-        for root, dirs, files in os.walk(a_data_folder.path, topdown=False):
+        for root, dirs, files in os.walk(os.path.join(config.DATASET_PATH, a_data_folder.path), topdown=False):
             if root == a_data_folder.path:
                 continue
             for file in files:
@@ -127,50 +143,59 @@ class backbone_dataset(Dataset):
         return  int((config.CACHE_SIZE_LIMIT / self.bytes_that_I_need) * self.total_number_of_frames)
 
     def print_memory_summary_table(self):
-        a_table_head = ["Memory Summary", "", "", "Limit"]
+        a_table_head = ["Memory Summary", "Size", "%", "Limit"]
         a_table = [
             ["CACHE that I will Use", f"{self.last_frame_in_cache*self.bytes_per_frame/1e9:.2f} GB", f"{self.last_frame_in_cache/self.total_number_of_frames*100:.2f} %", f"{config.CACHE_SIZE_LIMIT/1e9:.2f} GB"],
             ["SLOW MEMORY that I will Use", f"{(self.total_number_of_frames-self.last_frame_in_cache)*self.bytes_per_frame/1e9:.2f} GB", f"{(self.total_number_of_frames-self.last_frame_in_cache)/self.total_number_of_frames*100:.2f} %"], 
-            ["TOTAL that I Need", f"{self.bytes_per_frame*self.total_number_of_frames/1e9:.2f} GB", ""]
+            ["TOTAL that I Need", f"{self.bytes_per_frame*self.total_number_of_frames/1e9:.2f} GB", ""],
+            ["RAM that I will Use for storing data retrival info", f"{self.ram_for_stooring_data_retrival_information/1e6:.2f} MB", ""]
         ]
         print(tabulate(a_table, headers=a_table_head, tablefmt="grid"))
-    
-    def load_data_in_cache(self):
-        data_folder_index = None
-        frame_index = None
+
+    def make_data(self):
+        # Let's create the numpy arrays
+        self.data_path = np.zeros(shape=self.total_number_of_frames, dtype=np.dtype(('U', self.max_chars_data_folder_name_lenght))) # Unicode String containing exactly the needed ammount of chars
+        self.ram_for_stooring_data_retrival_information = self.data_path.size * self.data_path.itemsize
+        self.data_id = np.zeros(shape=self.total_number_of_frames, dtype=np.dtype('uint32')) # MAX id is 4.294.967.295
+        self.ram_for_stooring_data_retrival_information += self.data_id.size * self.data_id.itemsize
+        self.data_should_be_in_cache = np.zeros(shape=self.total_number_of_frames, dtype=np.dtype('?'))
+        self.ram_for_stooring_data_retrival_information += self.data_should_be_in_cache.size * self.data_should_be_in_cache.itemsize
+        self.data_already_in_cache = np.zeros(shape=self.total_number_of_frames, dtype=np.dtype('?'))
+        self.ram_for_stooring_data_retrival_information += self.data_already_in_cache.size * self.data_already_in_cache.itemsize
+        
+        # Let's fill them with data
         index = 0
-        pbar = tqdm(total=self.last_frame_in_cache+1)
-        quit = False
-        for i, data_folder in enumerate(self.data_folders):
+        for data_folder in self.data_folders:
             for ii in range(data_folder.elements):
-                images = []
-                for folder in config.DATASET_FOLDER_STRUCT:
-                    images.append(cv2.imencode(folder[1], cv2.imread(os.path.join(data_folder.path, folder[0], f"{ii}{folder[1]}"))))
-                self.cache[index] = images
-                pbar.update(1)
+                self.data_path[index] = data_folder.path
+                self.data_id[index] = ii
+                if index <= self.last_frame_in_cache:
+                    self.data_should_be_in_cache[index] = True
+                else:
+                    self.data_should_be_in_cache[index] = False
+                self.data_already_in_cache[index] = False
                 index += 1
-                if index > self.last_frame_in_cache:
-                    data_folder_index = i
-                    frame_index = ii
-                    quit = True
-                    break
-            if quit:
-                break
-        pbar.close()
-        pbar = tqdm(total=self.total_number_of_frames-self.last_frame_in_cache)
-        first_iteration = True
-        for data_folder in self.data_folders[data_folder_index:]:
-            if first_iteration:
-                start = frame_index
-            else:
-                start = 0
-            for ii in range(start, data_folder.elements):
-                self.slow_memory_path[index] = FramePath(data_folder.path, ii)
-                pbar.update(1)
-                index += 1
-            first_iteration = False
-        pbar.close()
+    
+    def __len__(self):
+        return self.total_number_of_frames
+    
+    def __getitem__(self, idx):
+        path = os.path.join(config.DATASET_PATH, self.data_path[idx])
+        id = self.data_id[idx]
+        data_should_be_in_cache = self.data_should_be_in_cache[idx]
+        data_already_in_cache = self.data_already_in_cache[idx]
+        if not data_already_in_cache:
+            data = tuple([cv2.imencode(data_folder[1], cv2.imread(os.path.join(config.DATASET_PATH, path, data_folder[0], f"{id}{data_folder[1]}"))) for data_folder in config.DATASET_FOLDER_STRUCT])
+            if data_should_be_in_cache:
+                self.cache[idx] = data
+                self.data_already_in_cache[idx] = True
+        else:
+            data = self.cache[idx]
+        return data
+
 
 if __name__ == "__main__":
     my_dataset = backbone_dataset()
+    for data in tqdm(my_dataset):
+        pass
     my_dataset.close()
